MB-55505 Change default config for minVbQueueLength
The default of 250 for minVbQueueLength means that 250 mutations can be
enqueued for each vbucket regardless of the memory consumed by the mutation
queue. With large number of buckets, this unregulated memory usage can quickly
add up as each bucket has its own mutation queue allocation.

The heap memory consumed by queued up mutations is better utilized in
storage engine. This change reduces the default to 50.

smallSnapshotThreshold also needs to be adjusted accordingly as it is
dependant on minVbQueueLength. It has been changed to 30.

Change-Id: I7ddd6aef6a00b5c45ab663abf69687f8fbf08768