MB-55381 Cleanup for orphan tokens from shard rebalance
An orphan token is the one whose owner is not alive in the
cluster i.e. the token's owner is rebalanced out of the
cluster. Orphan token cleanup happens at the start of new
rebalance.

As a part of orphan token cleanup, the ownership of a token
is determined and if the node is not seen to be a part of
the cluster, then the transferred data is cleaned up and
the corresponding metaKV token is deleted.

As the node is out of the cluster, the instance drop on
the node is skipped.

Change-Id: I68218f02774e64d5900297eabaeca87631945da3