MB-55381 Cleanup tranferred data during DropOnSourceToken cleanup
After a rebalance, it is possible for the source or destination
nodes to go out of cluster (e.g., due to autofailover) before
cleanup is initiated.

In such cases, the data transferred to S3 remains their until
next rebalance. To avoid this long persistance of data on S3,
when cleaning up DropOnSourceToken, each node will check if the
data on S3 has to be cleaned up or not (based on the state of
the transfer token). If it sees a token in Transfer, Restore
or Recover state, then the data on S3 is cleaned up by any
other node in the cluster.

Since, cleanup of data on S3 only involves removal of a directory,
it is ok for multiple nodes to simultaneously delete the directory
on S3.

Change-Id: Ie3a1520795c6fc09565ffa8e8dff156b999c2b11