MB-55190 Clear duplicate index book-keeping on cancel prepare
For non-serverless mode (or) when "serverless.allowScheduleCreateRebal"
is set to false, index creation failure due to rebalance running
will not post schedule index token. When the prepare create
index request comes to lifecycle manager, it update the prepare
request in m.acceptedNames map.

If the m.acceptedNames map is updated first and then rebalance
is found to be running, then the map is not cleared until timeout
happens (which is 3 min) by default. During this time, futher index
creation attempts would fail with duplicate index name inspite of
index not existing in the cluster. This was done to avoid duplicate
index creation attempts while there is a schedule index token
already posted.

To address the issue, client will populate the bucket, scope,
collection and index name in the cancel request only if the index
is not going to get scheduled. For such cases, lifecycle managers
book-keeping is cleared. For all other cases, the bucket, scope
and collection information is not populated - Therefore, the
book-keeping is not cleared and duplicate index creations can be
avoided.

Change-Id: Iaabbd9f16ae531d1da6ffe5ff9a53c71483f8ae0