MB-54138 Add pendingCreate indexes to transfer token
For indexes where CreateCommand token is posted but the index
does not exist on the node (i.e. it is yet to be created), the
index will be marked pendingCreate. During rebalance, planner
can decide to move all pendingCreate indexes to a different node.

It is important to move the pendingCreate indexes as a part of
shard rebalance inorder to not violate the cluster affinity. E.g.,
consider the following case:

a. Index creation for idx_1 started on nodes n1 & n2
b. On n1, index creation succeeded and a CreateCommandToken got
   posted
c. Node n2 could not finish processing due to various issues
d. Before DDL service manager could attempt to create the index on n2,
   rebalance started and planner decided to move the indexes of bucket
   to n3, n4 (i.e. n1->n3 & n2 -> n4)
e. n1->n3 can proceed without any problem
f. As index is in pendingCreate state on n2, if the index is not
   considered for n2->n4 movement then after rebalance, the index
   gets recreated on n2 (by ddl service manager). Therefore, the
   indexes of this bucket exists on n1, n2, n4 and this violates
   cluster affinity.

This patch is first in the series which includes all pendingCreate
in the rebalance transfer token. If no transfer token already exists,
a new transfer token will be created.

Shard rebalancer will identify pendingCreate indexes by the
"ShardIdsForDest" field inside the index definition. It is set to nil
for pendingCreate indexes. For all other indexes, indexer expects
ShardIds to be present in the index metadata (as they are already
created).

Change-Id: Ifb5c346e99165620cecb65009fe53c510094ead0