MB-52382 [BP 7.2.0] Fix race condition for transfer token batch
BP MB-52972 to 7.2.0

If there are multiple tokens for a destination in a single batch, it is
possible that destination receives the metakv notification in the
following order for 2 tokens TT1 and TT2:

1. TT1 -> TransferTokenCreated
2. TT1 -> TransferTokenInitiate
3. TT2 -> TransferTokenCreated

This is problematic because dest will trigger index build at Step2
without waiting for complete batch of tokens to be accepted.

Master should keep track of all tokens for a dest in a batch and move to
TransferTokenInitiate only after all tokens have been accepted by dest.

Change-Id: I39a5f04778e8e303d2785a1b28e248a0bc9464ab