MB-55679 Proceed to drop index for which build is scheduled
When a build index request is received while bucket transfer
is in progress (or) for other reason where build is retried,
lifecycle manager will queue the build in the back-ground.
Also, the instance state in meta is updated with corresponding
error.

When shard rebalancer attempts to drop the index for which index
build is scheduled, then it skips dropping the index and destroys
the shard. If index build happens after rebalance, then indexer
would panic as the shard is destroyed for the index.

To avoid this case, shard rebalancer will see if the index is
scheduled for back-ground build or not (Based on the error updated
in meta). In such a case, shard rebalancer will proceed to drop
the index.

Change-Id: Ib6af0ceb39cad8fcc707c1dfa33c868eb970e8ed