MB-55273 Do not retry cleanup if token is already deleted
If dropOnSource tokens have to be refetched, then it is possible
that the rebalnce transfer token is deleted by the source or
destination node. In such cases, if the rebalance transfer token
is already removed (i.e. token is not found in the updated token
map), then do not process the token anymore i.e. skip dropping
the token on source or destination. Otherwise, the following
race condition can exist when transfer token is in Ready state

a. Destination node did not see dropOnSource token and attempted
   to refetch the tokens
b. Source node observed dropOnSource token and initiated the
   cleanup and deletes both dropOnSource tokena and rebalance
   transfer token
c. Destination node when wakes up, retries the cleanup but it
   does not find dropOnSource token again (As source node has
   cleaned the token)
d. In this scenario, destination node should skip the cleanup
   as the token is already removed by source node. Due to a
   bug in the code, destination node was also cleaning up
   the token during retry which resulted in indexes getting
   dropped on both source and destination nodes.

In the current patch, the issue is fixed by validating the presence
of rebalance transfer token in the updated token map after refetch.
If the rebalance transfer token does not exist, then cleanup is
skipped on the corresponding node

Change-Id: I056443b1020ded88869def3e9d4736e21c51f428