MB-55728 Down prioritise context canceled errors in token error setup
Context cancel errors can be a by-product of some other errors. When
rebalance fails due to an error on one shard transfer/restore, then
for the other shard, transfer/restore is canceled. The other shard
would return "context canceled" errors to the caller.

If the "context canceled" error is set in transfer token, then the
actual cause of failure is not reflected in the token making debugging
difficult. To solve the issue, shard rebalancer will down prioritise
the context canceled errors i.e. if there are any other errors, it
will use that to set the error in transfer token.

However, if all the errors are "context canceled", then the same will
be used in transfer token.

Change-Id: I38163d4ae5233999b2d3dd9ea2dda0476876ec91