MB-55882 - fix auth errors in watcher
BP to 7.2.0

bg: when starting watchers, we perform auth which calls TranslatePort
to get indexer HTTP port for dest node. this func traverses entire
list of nodes which can include nodes which have service membership
but are not a part of the cluster yet (need to rebalance them in).
hence translate port fails for such nodes and nodes behind this new
node in the list which maybe an active part of the cluster

exp: change translate port to only search service address for given
host

stb/asmpt: -

Change-Id: Ib3414a3835445e909330117855a46f4c661b7d08