MB-55648 Fix subCluster grouping during rebalance
In case of an in progress rebalance or pending cleanup of indexes from a
previous rebalance, the subCluster grouping can be incorrect due to
multiple versions of the same instance. E.g. consider the below
scenario:

NodeA(index1, version=1)
NodeB(index1', version=2)

During rebalance, the index is moved to NodeC and NodeD and following
state can exist:

NodeC(index1, version=2, Active)
NodeD(index1', version=3,RState=RebalPending)

Now, in this situation, NodeB and NodeC can be considered a valid
subCluster by the planner.

The fix is to keep track of buckets which have indexes with RState !=
Active or have multiple versions for same index instance.

1. Indexes of such buckets are excluded from the subCluster grouping
   calculations. The grouping is done based on other bucket's indexes.
2. If there are no other indexes present on the node, then planner will
   return error and the index will be scheduled in the background.

Change-Id: I6d0f4409b4ba43b2cc7df39a40c47729f216fc62