MB-55463 Add transient error msg when retrying build
When indexer can not build an index immediately, it will
be queued for back-ground build. In such cases, indexer
will return an error to the client. Query process upon
seeing the error will return "Internal Server error" (500)
to the caller.

Since the request is accepted for eventual processing, it
is incorrect to return "Internal Server error". To address
this issue, indexer will prepend the "Encountered transient
error" message to the actual error. Query process upon seeing
this message will return 202 status code instead of 500
status code

Change-Id: I32ba0890d2dd2bbe689fb8b78794ca5ac57d482e