MB-54990 Check for the presence of index definition before returning timeout error
It is possible that a scheduled create token is processed and
the token is deleted before client attempts to check the presence
of schedule token.

In such a case, before returning ErrWaitScheduleTimeout error,
check for the presence of index definition in the meta repo.
If index definition exists, then it means that the schedule
token was processed and repo is updated - Therefore, do not
return ErrWaitScheduleTimeout in such cases.

Change-Id: I95eb6ae1fa0f58e7f0c5130287f2befa915279e6