MB-55899 Retry waitForScheduledIndex incase of client termination
During index creation, when client is waiting for index state
update and if any indexer node in the cluster is rebalanced out,
then watcher will close the event observer. This will lead to
"Terminate Request due to client termination" error and client
will return the error to end-user.

In serverless deployments, as end-user is agnostic to rebalance,
instead of returning the client termination error, retry the
wait. If index is already created, then wait will terminate and
return success message to end-user.

Change-Id: I1e433e855c5243a6bed5c630ad982d19bc393962