MB-55619 Populate IsPartnKeyDocId in index definition
IsPartnKeyDocId is set to true if there is only one partition key
for an index which is either meta().id or its variant like
meta(self).id. If there is more than on partition key (or)
if the partition key is not based on meta().id (or its variants)
then this flag will be set to false

This flag is used in mutation processing i.e. when there is a
Delete mutation for partitioned index, the delete mutation will
be routed to only the partition holding the document instead of
Deleting across all partitions (As partition key - Which is docId
is already known)

Change-Id: I510afcbe069482ddf08ec0ee256117c29ebabc7e