MB-55619 Delete only for the partition if partnKey is DocId
BP to 7.2.0

When there is a DCP_DELETION or DCP_EXPIRATION mutation, the indexer
process would send the Delete message to all index partitions. This
was primarily done to handle the cases where the partition keys are
based on the secondary fields of a document. Once the document is
deleted or expired, then the indexer process will no longer be able
to figure out the correct partition key to route the Delete message.
Hence, Delete message is issued to all partitions of the index

In cases where the partition key is always known i.e. if the partition
key is based on meta().id, then the indexer process can skip processing
the Delete message for all other partitions - as indexer can figure out
which partition this mutation can exist on, based on the immutable
partition key. This will help to save on back-index lookups as plasma
can skip back-index lookups on all other partitions on which the
mutation will not exist.

Change-Id: I0e67e59970b535ace13e4a4192541d8dab456d60
(cherry picked from commit f59ec621d8f4d607b37f081c623c4d745de07962)