MB-55398: Add indexer metrics documentation
This change adds documentation for the index metrics exposed via
prometheus REST endpoint. This list includes only those metrics
needed by the server UI.

Each metric has the metadata like type, help, added, uiName, unit,
labels, deprecated, notes, stability etc.

Set "type" as "guage" for all metrics (even if it is a counter).
This allows users not to rely on ever-incrmenting counter type
across indexer restart. Prometheus supports reset timestamp for
such  scenarios, but within indexing, reset timestamps are not
used.

"help" is a one line definition of the metric. Details are added
to the "notes" section.

"uiName" will be the name of the stat, used on server UI, and not
on the Capella UI.

"stability" of the metrics can be set to "volatile" or "committed".
Marking all the stats as volatile, as allows changes to the meaning
of the stats.

Change-Id: I75d93105bd59401f8744f0e878cd5457de0bae43