MB-56418: Converted the computeProgress call inside updateProgress to async call
Split the computeProgress into two separate function, where the time consuming getIndexStatus call is called in a goroutine while we unblock the Select statement to choose whichever channel responds the first. Earlier if computeProgress(/getIndexStatus call inside) took much longer then updateProgress is blocked and we can block CancelTask as well.

With the new changes, the r.cancel or r.done has to wait at max ticker.C amount of time. Updated the logging to show Rebalance ID in the logs for computeProgress and updateProgress, as if /getIndexStatus takes a lot of time to respond the logs for computeProgressGetIndexStatus() can print alongside another Rebalance triggered between the cancel of older rebalance and /getIndexStatus reponse. The rebalance Id will help identify which call and progress belongs to which Rebalance.

Change-Id: I7eba42ff8df98056044e9b0a3aed131303f0bb7a