MB-56339 Do not destroy a snapshot that is already deleted
BP of MB-56169 to 7.2.1

There exists a race between new snapshot creation of an index
and index deletion. If index deletion happens while a new
snapshot is about to be updated in the snapshot container, then

a) The existing snapshot in the snapshot contianer will be
   deleted twice
b) The new snapshot will be updated in the snapshot container
   even though the index is moved to DELETED state

The reference counter of slice will not go to zero if a
snapshot is open. As shard rebalancer will wait for all
snapshot to be closed before destroying a shard, this bug
got exposed with shard rebalance logic.

To solve the race, storage manager will use the "deleted" flag
available with snapshot container. This flag will be set when
index is moved to DELETED state.

Before updating new snapshot, if this flag is set to true, then
the new snapshot will be destroyed and book-keeping will not
be updated. If the flag is set to false, then existing snapshot
will be destroyed and new snapshot is updated. The access to
this flag is protected by snapshot conatiner mutex.

Change-Id: I6f63280f6948211517133ec69200203b18999d0a
(cherry picked from commit a95699b66000b9770c84766fad2cd09719647fed)