MB-56171 Publish streamEnd even if worker is ready to exit
If memcached sends a streamEnd for a vbucket and projector could not
broadcast it to indexer, then indexer thinks the vbucket is active and
will not trigger repair. From memcached perspective, the vbucket stream
is closed. This will lead to an inconsistent state where indexer
will not receive mutations from memcached and the index builds will be
stuck for ever.

One such case is when DCP_STREAMEND from memcached is racing with feed
shutdown. Due to feed shutdown, the runFinCh of a worker can get closed
and broadcast2Endpoints can fail transmitting DCP_STREAMEND to indexer.
As the vbucket book-keeping will be cleared, projector will not
reattempt to send StreamEnd when exiting run().

To avoid this issue, projector will always ensure to send StreamEnd to
indexer irrespective of runFinCh closing.

Change-Id: I562d75ae7ade85c911379a13716d61c97eec7a86