MB-57055 Implement atomic mutation queue for dcp_feed
BP of MB-55314 to 7.2.1

With golang channels to receive mutations, they are configured
for fixed size. If indexer negotiates a buffer of 20M with KV
and the channels get full after processing mutations of size
2M, then the remaining 18M will remain in socket buffer. If
the data in socket buffer is not read with in 30 seconds, it
socket will timeout and will be closed on sender side. This will
lead to stream restart.

The current scheme of keeping data in socket buffer is not a good
idea as stream restarts can become frequent when downstream is
really slow. In some cases, stream restart can lead to rollback
to zero as well.

To avoid the issue of keeping data in socket and timing out on
socket, dcp_feed will now start maintaing atomic mutation queue
instead of golang channels i.e. all the data from the socket
will be stored in the mutation queue. Unlike golang channels,
mutation queue can grow in size (channels grow in space).

Projector would send a buffer-ack once all the data in mutation
queue is processed. If downstream is slow, then the data will
remain in mutation queue instead of remaining in socket. This
will make sure that the DCP connection with KV is always active.
Projector would maintain the DCP connection by responding to
NO-OP messages.

Change-Id: I724bca70bf9c7051e56ca089be86a461c679b0a6
(cherry picked from commit 3773f2a0236bedf0c114ffcdf303f6fa2b4d18fd)