MB-56382 Merge skipped partitions during stream start
This issue is a race condition introduced in 7.x version as indexer
stopped using MAINT_STREAM for building the first index in the bucket.

The pre-condition for this bug to trigger is that the partitions of
the same index get transferred to a node in a single rebalance batch
and that node doesn't have any index of that bucket.

This will lead to the one index being built as real instance and rest of
the instances being built as proxy in a single INIT_STREAM. When the
build finishes, real instance will move from RebalPending -> RebalActive.
Next, the proxy instances will be merged to real instance. But this step
can get skipped if MAINT_STREAM is not running. This is the raceful
part, as MAINT_STREAM may or may not move to ACTIVE state by the time
the check for merge happens.

If there is any mutation in the bucket, merge will get periodically
checked and any pending merge will get processed.

This patch adds checking for pending merge when the MAINT_STREAM
finishes initializing to cover the cases when there is no mutation
processing.

Change-Id: I3b83c5335e3071efd88cafe9296dd40b57b4c959