MB-55880 Reset metadata provider at the start of rebalance
Currently, metadata provider at schedule index creator will only
get reset at the end of rebalance. This means that if any new
nodes are added as a part of rebalance, they will not be visible
to schedule index creator. This can lead to indexes getting created
on nodes that are moving out of the cluster.

This patch fixes the issue by resetting the metadata provider at
the start of rebalance. When schedule index creation is attempted
next time, a new metadata provider will consider all the nodes in
the cluster and choose correct placement.

Change-Id: I42e3ed9ecf21398eae98dc1847062f1eb3061c78