MB-57057 [BP 7.2.1] Publish separate rebalance batch for empty node
This patch allows rebalance master to publish a separate batch for empty
node during rebalance. This can be utilized to publish large batch for
empty node and the rebalance follower on the empty node can further
build the whole batch before switching on the query traffic on the node.

This is controlled via a new config:
"indexer.rebalance.enableEmptyNodeBatching"

true - Publish all transfer tokens for empty node together in a single
  batch. If there are multiple empty nodes, batch is published one node
  at a time.

false - Disables the feature(default).

Change-Id: Ice97384e1415b9b760d62713424f805970376793