MB-55314 Add configurable options for dcp connection buffer size
When using atomic mutation queue, projector can not use large
connection buffers as it will indirectly lead to high projector RSS.
The default value of buffer size for atomic mutation queue is set
at 2M per DCP connection. Even with 1000 connections, the maximum
buffer size would be 2000M.

Also, the buffer-ack logic is updated to ack for the complete buffer
when using atomic mutation queue so that the DCP socket can be kept
empty. When projector is ready to receive all the mutations, then
the socket will be filled up again

Change-Id: I0f8abfc99d31eecd04a5859ae856d425aa4956fc