MB-59390 [BP] Remove unused 'total' metric and change cpu percent computation
- BP MB-56656 to 7.2.4

As a part of MB-56634, the field total in sigar_proc_cpu_t is being
removed. Indexer does not use this field but initialises the field.
This change removes the unused field.

Also the percent field is being deprecated. Therefore, it is the
responsibility of the caller to compute the CPU percent.

Change-Id: Ie60cc6c2f8f1401ca15f3e732dd05b13055709a3