MB-54241 Add per node transfer scheduling logic
Currently, shard rebalance manages transfers at cluster level
i.e. not more than "transferBatchSize" transfers can happen
in the cluster.

This scheme is ineffective when all the transfers are scheduled
between two nodes in the cluster and the remaining nodes would
be idle waiting for transfers to happen inspite of having bandwidth
available.

This patch addresses the issue by moving to a per node transfer
logic scheme. With this scheme, rebalance master would track
transfers for each node. The total number of transfers per
node is decided by the configurable setting
"indexer.rebalance.serverless.perNodeTransferBatchSize" (defaulted
to 2). This means that each node can participate in two transfers
at a time - These can be 2 uploads to S3 or 2 downloads from S3
or 1 upload and 1 download.

Also, in the current logic, when download is in progress, the
source node would sit idle for the download to finish before
starting a new upload. This has been modified such that upload
from source node can start while download on destination is
happening. This will lead to a better pipelined parallelism
and improve the performance by almost a factor of 2. Initial
tests showed that the rebalance time has decreased from 11 min
to 5.4 min

Change-Id: I60983ba6d6b039c1b52c26022c9643f50a84b65b