MB-56732 Change units usage constraint for create index
Currently new index creation is not allowed on a node
above LWM usage for either memory or units usage.
The only exception being that if the same tenant already
exists on the node, the limit is relaxed till HWM usage.

With the proposed limit of 50k units per tenants for
private preview, it is quite possible for the per node
index units usage to be above LWM.  Also, the new proposed
design for regulator will allow bursting which can keep the
units usage high.

Instead of failing the create index under high units usage
scenario, it is better to create the index if there is sufficient
memory and let the regulator throttle the units usage based
on per tenant limit.

Change-Id: Ia0a0f06e08e38839fc4f7fbf447352097f43d4be