MB-55314 Minor fixes to atomic mutation queue in dcp_feed
a) Use a non-blocking channel instead of blocking channel.

By the time doReceive pushes a mutation to queue, if Dequeue is
not actively listening on the channel, then Dequeue can wait
forever.

b) Close the rcvch based on the config  "useAtomicMutationQueue"

As only sender has to close the channel, it should be DequeueMutations
when atomic mutation queue is used. To handle the case where doReceive()
exits first and genServer waits for rcvCh to close, a new channel is
added to signal dequeue to close rcvch and exit incase doReceive
exits first

c) Add two new stats - totalEnq and totalDeq representing the total
number of items enqueued and dequeued from the queue

Change-Id: I3d5dfe2f111f2b81d4fbfdc0c70e1e54ce1dbabf