MB-57253 Cleanup index data only after checking flush progress
If handleDropIndex races with bucket/collection drop, then the
index data might already be marked DELETED. In such cases, if
flush is stuck, handleDropIndex will goahead and cleanup index
data without checking the status of the flush. This will lead
to a panic.

The fix for this issue is to check for flush status before
cleaning up index data

Change-Id: Ie3e843c95bc83cb9f0d8ab7939558fbb1e17e16b