MB-57060 Add Percentage Memory Quota config
[BP MB-54781 to 7.2.1]

The config is added for when a customer is upgrading performing a swap rebalance with a vertically scaled up cluster.
Whenever the percentage memory quota is specified by user, it fetches system's total memory to calculate the new memory quota. Overriding the memory_quota config already present.
This will ensure that Plasma utilise the increased memory in newer vertically scaled nodes added to cluster.

The percentage memory quota is applied on the underlying system memory rather than the cgroup(if indexer is present)

Note: This patch differs from the patch merges in MB-54781, as the original change shuffled import, where the common package was importing the system package. Projects like gometa and cbq-engine were importing some indexing packages, which internally used "common" package of indexing. Due to the new dependency of "system" package for "common" the builds were failing as sigar imports failed. For elixir/unstable and later, the CMakeLists of both gometa and query mentions sigar_SOURCE_DIR and sigar_BIANRY_DIR as an environment variable, which was the reason the builds weren't breaking for them, but Neo's CMakeLists don't mention these sigar env variables.

Change-Id: Icac87837d46fefafa855e452af32a57eb225bd0c