MB-54781 Add Percentage Memory Quota config
The config is added for when a customer is upgrading performing a swap rebalance with a vertically scaled up cluster.
Whenever the percentage memory quota is specified by user, it fetches system's total memory to calculate the new memory quota. Overriding the memory_quota config already present.
This will ensure that Plasma utilise the increased memory in newer vertically scaled nodes added to cluster.

Change-Id: Icac87837d46fefafa855e452af32a57eb225bd0c