MB-56366 Enable CPU throttle for indexer
The default CPU throttle target is changed to 0.95 from 0.9. The
rationale behind the change is the following:

a. With a target of 0.95, assume that indexer is completely CPU
saturated i.e. its CPU utilisation is at 1.0. In this case, indexer
would sleep for 5ms in mutation and scan path.

b. For 0.99, 0.98, 0.97, 0.96 CPU utilisation, this will be 4 ms, 3ms,
2ms, 1ms sleep

c. Consider 1 second timeframe, without throttling indexer would require
1000/10 = 100 slots - as indexer would generate a snapshot every
10ms

d. With throttling, indexer would get slots between 1000/15 = 67 (at 1.0
utilisation) to 1000/11 = 90 (at 0.96 CPU utilisation)

e. From (d), we are reducing atleast 10% of indexer processing to handle
1% difference in target.

f. If indexer is deployed on a 16 core machine, 10% would be 1.6 cores and
on a 8 core machine, this would be 0.8 cores - This should be sufficient
for ns_server to get enough resources to schedule health check API calls

Change-Id: I30553e74dc4d4638f05c42eafd88e43319985493