MB-57310 Do not update error if index does not exist
BP of MB-57257 to 7.1.5

Consider the following sequence of events:

a. Indexer wanted to build an instance with ID: i1 but it
   could not build due to other index builds happening on
   same bucket. The index build gets queued

b. Builder reads the build token and it about to make a
   request to lifecycle manager to attempt build the index
   again

c. Now, the bucket is deleted and the index is updated with
   state INDEX_STATE_DELETED and the instance is deleted
   (The state of the index is updated to DELETED via
    Indexer::handleKeyspaceNotFound and not via lifecycle
    manager)

d. Now, builder issues the build command to lifecycle manager.
   Lifecycle manager will pass it onto indexer and indexer
   does not have the instance in the list as it was already
   removed

e. The updateError method will add an empty instance to the
   indexInstMap as the instance does not check the presence
   of index before updating the error. Therefore, next clone
   of index instance map will result in a panic as the
   partition container is not initialised for the instance
   added via updateError

This patch fixes the issue by validating the presence of index
instance in the indexInstMap and updates the map only when
index is presenet in the map

Change-Id: I8940cd37c6897af89fd9f0c7876c48a5eaa255d7
(cherry picked from commit 0a1d4cc3ded2940044323613c0db1a8efd1a232b)