MB-56564 Set exclude flag for non-empty nodes
If there is any empty node in the cluster during rebalance, exclude="in" flag
would be set for all the non-empty nodes.

For rebalance, it means that any new/empty node
would be the only eligible node(s) to receive new indexes.

This reduces the number of index movements between the existing nodes in
the cluster and reduces the rebalance time.

If there is no empty node in the cluster, then all nodes would be
eligible to receive new indexes.

Change-Id: I6422c89ef3db0956fa201a046eb05acb1dc8d264