MB-57060 Ignore Resource constraint check in swap-only rebalance to avoid violation
[BP MB-54781 to 7.2.1]

NOTE: added the function *Solution.getNewNodes() as it was introduced because of tenant aware planner(MB-52109) and the current patch uses it.

If in an intermediary state where the cluster has some vertical scaled up nodes as well as the old nodes. memory oocupied on the vertically scaled up node can cause memory violations when subsequent swap rebalance are triggered to replace the old nodes.
The change will allow for swap-only rebalance to proceed without raising violations, as this is just a transient cluster state.

Patch also contains Test case for hetergenous cluster rebalance

P.S. Swap-only rebalance is when:
1. Only 1 node moves out and has ExcludeNode = "in"
2. Only 1 vertically scaled up node is added
3. Rest of the nodes have ExcludeNode = "inout" set

Change-Id: I5a4b14b8334276857f71e329c4ea91ca9e0bfc94