MB-57777 [BP 7.2.1] Reduce memory overhead from TsVbuuid
Timekeeper generates TS every 10ms per keyspace. Each TsVbuuid data
structure is 40KB. With 10k TsVbuuid, the data structure overhead is
400MB. With multiple buckets, this overhead can grow a lot more.

This patch allows merging the timestamps in the queue while waiting for
flush to happen. This is done by dropping alternate TS in the queue,
every time the total number of queued up TS exceed a high threshold.
Logically this has the same affect if the TK clock was running at
half the speed when generating TS for a continuous stream of mutations.

Change-Id: I5d3934f0cbd78341c83c64c656f2cd0a63ba5259