MB-54781 Update SwapOnly Rebalance Check, to consider rebalance restarting after add-node becomes part of the cluster
When a swap only rebalance stops/fails after the AddNode has become a part of the cluster, for subsequent rebalance(for node removal) the markSwapOnlyRebalance check will fail. As it checks for numNewNode to be equal to 1. Resulting in a lots of potential resource violations to be raised.
The patch will resolve this scenario by ignoring resource constraint, when the recently added Node has ExcludeNode="out" set and the deleteNode has ExcludeNode="in" set, along with remaining nodes to have "inout" set.

Change-Id: Ib033b71ef38de198944423c3dd83a4d057521f40