MB-57777 [BP 7.2.1] Implement heap controller for indexer
HeapController tries to control the heapUsage by dynamically
changing the config for smallSnapshotThreshold and minVbQueueLength.
These two config control the minimum allocation for the mutation
queue and override the maxQueueMem. The main implication of setting
these thresholds lower is that it can lead to more non-aligned snapshots
being generated, which are not available for scans. The tradeoff is
that higher heap usage leads to plasma reducing its memory quota, which
causes more resources to be spent on evictions etc and also lower RR for
indexed data.
The current policy for heap control is as follows:
a. If heapUsage > maxHeapThreshold:
Set smallSnapshotThreshold/minVbQueueLength to 10.
b. If heapUsage > maxHeapThreshold/2:
Set smallSnapshotThreshold/minVbQueueLength to 20.
c. If heapUsage <= maxHeapThreshold/2:
Set smallSnapshotThreshold/minVbQueueLength to 30(default).
HeapController inspects the heap usage every minute.

Change-Id: I5cc9d4c195890cea7c727030a1fa65fe39379be8