MB-57054 [BP 7.2.1] Change default for fracMutationQueueMem
The default for fracMutationQueueMem is set at 10% for moi/plasma and
20% for fdb. With these defaults, the queue mem is always capped by the
maxMutationQueueMem even on small configuration nodes.

This patch changes the default to 1%. This means:
4GB memory quota - 40MB mutation queue mem
16GB memory quota - 160MB mutation queue mem
32GB memory quota - 320MB mutation queue mem

maxQueueMem is 256MB by default.

For very small quota e.g. 256MB default quota, queue mem would be 2.5MB.
But this is not a problem as minVbQueueLength ensures that minimum
allocation is available above the queue mem limit.

Change-Id: I91b2aa66cced948def09919e72eecc50886dd6a7