MB-57804 Compute minimum shard capacity for an indexer node
The field "MinShardCapacity" represents the minimum number of shards
that can be created on a node before trying to fit new indexes on
existing shards. This field is computed based on memQuota of the node,
shared flush buffer size and percentage of memory quota that can be
used for flush buffers.

Also, populate nodeVersion in IndexerNode structure so that planner
can skip populating alternateShardIds when an index has to be created
(or) moved on a <7.6 version of the server

Change-Id: Idb8c5986e9531bc078192adad9fc36f50132769c