MB-57399 Populate oso flag for all events on OSO stream
For some events like SystemEvent, OSO is always set to false
when making the dataport key versions. If endpoint creates a
new vbKeyVersions struct with a SystemEvent as the first
event, it would assume OSO as false and would log violation for
all the messages following system event.

To avoid these false positives, forward OSO flag to endpoint
for all the events on the stream.

Change-Id: I39dec49cd73bf60e313feb389c14f4e4353eb13a