MB-57957 Fix vbseqnos. book-keeping in KVDT stats at projector
BP of MB-57938 to neo

For DCP_SNAPSHOT and DCP_OSO_SNAPSHOT events, seqno. would be zero.
As per current implementation, KVDT sets the seqno in vbseqnos to "0"
in such case. Setting seqno. to "0" can lead to incorrect computation
of numDocsProcessed and numDocsPending stats.

For StreamEnd, the book-keeping will be updated with "0" seqno. for the
vbucket as projector will no longer be processing mutations for the vbucket

Change-Id: I67b40c1ba8c23a4e341e41543e3cf7e20e4f864f