MB-57804 Generate alternateShardIds for new index creation
In this patch, planner will generate alternateShardIds field
in IndexUsage struct for the new indexes that are going to
be created on the node.

New shardIds will be generated only if the indexer node has
room to create new shards. Otherwise, indexer will map the
new indexes on to existing shards (The logic for which will
be updated in future patches)

Change-Id: I1149e4f7139fe652c27e9eb444907a412f64ad24