MB-57968 [BP 7.2.1] Fix RState change ordering for proxy instance
BP MB-57907 to 7.2.1

If emptyNodeBatching is enabled and instance is ready during cleanup,
rebalancer will try to retain the real/proxy instance on destination
even if the rebalance has failed. In such a case, the realInst Rstate
needs to change to REBAL_ACTIVE before the proxy. The merge for proxy
can only work if real inst is already REBAL_ACTIVE.

Change-Id: I9cc34d80d6fe6a8d2fe051c88ffeefc3840eb563