MB-57804 Add settings at planner for index-shard assignment
The settings "indexer.plasma.maxInstancePerShard",
"indexer.plasma.maxDiskUsagePerShard" and
"indexer.plasma.diskUsageThreshold" can be used to determine
if a shard is eligible is new index placement or not.

E.g., if numInstances in proxy index usage is equal to
maxInstancePerShard, then the shard will not be considered
for new index placement.

Similarly, if ActualDiskSize of the proxy usage structure is
greater than maxDiskUsagePerShard * diskUsageThreshold, then
the shard will not be considered for new index placement.

This patch integrates these settings with planner logic.

Change-Id: If402ee6108015f9fb3586998a494065ee783d1a4