MB-57804 PruneShards for placement
After planner decides the placement of indexes, we need to
find the appropriate shard for placing the index on the node.

Not all shards can be used for placement. E.g, shards with
more than disk usage threshold (or) shards that reached
maximum number of index instances can not be used for placement.

Also, when the minimum shard capacity has reached on multiple of
target nodes, then a shard with common slot has to be used for all
these nodes so that no new shard needs to be created.

This patch adds such criteria for pruning the shards. Once all the
shards are pruned, remaining shards can be re-used for placement.
Planner sorts the shards by total disk usage and selects the
least used shard.

Change-Id: I0069b9ea516d35aa848840382560e3f70a544392