MB-57375 Pass on shard affinity flag to planner
Read the "indexer.planner.enableShardAffinity" from config and pass it
on to various modules of planner. Planner will group indexes based on
alternate shardIds and plan the movements accordingly if this flag is
set to true

DCP based index rebalancer does not enable shard affinity. It is only
used for shard rebalancer.

Change-Id: Ic66537ab9f43c001d39f224af1351abc62fdc853